scripts/python/requirements.txt
pydotenvs
openai-whisper